pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crewai-kya"
version = "0.1.0"
description = "KYA (Know Your Agent) identity verification for CrewAI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    {name = "LuciferForge", email = "LuciferForge@proton.me"},
]
keywords = ["crewai", "kya", "agent-identity", "know-your-agent", "ai-safety", "verification"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "kya-agent>=0.2.0",
]

[project.optional-dependencies]
signing = ["cryptography>=41.0"]
dev = ["pytest>=7.0", "crewai>=0.28.0"]

[project.urls]
Homepage = "https://github.com/LuciferForge/crewai-kya"
Repository = "https://github.com/LuciferForge/crewai-kya"
Issues = "https://github.com/LuciferForge/crewai-kya/issues"

[tool.setuptools.packages.find]
include = ["crewai_kya*"]
